backend/requirements.txt
# Web scraping (scrape.py)
beautifulsoup4==4.12.3
aiohttp==3.13.3
tqdm==4.66.5

# Translation (translate.py)
requests==2.32.4
google-cloud-translate==3.16.0

# ML/AI (model.py)
torch>=2.9.0
openai==1.58.1
sentence-transformers==3.3.0
pinecone==5.4.2

# Web API (main.py)
fastapi==0.115.5
uvicorn==0.32.0

# Environment variables
python-dotenv==1.0.1
